separate api modules by access scopes
